+++ zerver/views/archive.py
@@ -6,10 +6,10 @@ from django.template import loader
 from zerver.lib.streams import get_stream_by_id
 
 from zerver.models import Message, UserProfile, get_stream_recipient
-from zerver.lib.actions import get_topic_history_for_web_public_stream
 from zerver.lib.avatar import get_gravatar_url
 from zerver.lib.response import json_success
 from zerver.lib.timestamp import datetime_to_timestamp
+from zerver.lib.topic import get_topic_history_for_web_public_stream
 from zerver.lib.exceptions import JsonableError
 
 def archive(request: HttpRequest,
